+++ addOns/ascanrules/src/test/java/org/zaproxy/zap/extension/ascanrules/CrossSiteScriptingScanRuleUnitTest.java
@@ -963,6 +963,40 @@ class CrossSiteScriptingScanRuleUnitTest extends ActiveScannerTest<CrossSiteScri
     }
 
     @Test
+    void shouldNotReportXssInUrlAttributeWhenPayloadIsModified()
+            throws NullPointerException, IOException {
+        String test = "/shouldNotReportXssInUrlAttributeWhenPayloadIsModified/";
+
+        this.nano.addHandler(
+                new NanoServerHandler(test) {
+                    @Override
+                    protected Response serve(IHTTPSession session) {
+                        String name = getFirstParamValue(session, "name");
+                        String response;
+                        if (name != null) {
+                            // Strip out < and >
+                            name = name.replaceAll("<", "").replaceAll(">", "");
+                            response =
+                                    getHtml(
+                                            "InputInLinkHrefTag.html",
+                                            new String[][] {{"name", name}});
+                        } else {
+                            response = getHtml("NoInput.html");
+                        }
+                        return newFixedLengthResponse(response);
+                    }
+                });
+
+        HttpMessage msg = this.getHttpMessage(test + "?name=file.html");
+
+        this.rule.init(msg, this.parent);
+
+        this.rule.scan();
+
+        assertThat(alertsRaised.size(), equalTo(0));
+    }
+
+    @Test
     void shouldReportXssInScriptIdTag() throws NullPointerException, IOException {
         String test = "/shouldReportXssInScriptIdTag/";